refactor(#152) :: 먹이급여량 int를 float형으로 변경

+++ src/main/java/com/command/toyvillage_server/domain/app/feed_log/presentation/dto/response/FeedLogAdminDetailsQueryResponse.java
@@ -12,7 +12,7 @@ public record FeedLogAdminDetailsQueryResponse(
         String animalKind,
         String animalName,
         String feedType,
-        Integer feedAmount,
+        Float feedAmount,
         LocalDateTime feedDateTime,
         String significant
 ) {
