+++ src/main/java/com/command/toyvillage_server/domain/app/feed_log/presentation/dto/response/FeedLogQueryResponse.java
@@ -20,7 +20,7 @@ public record FeedLogQueryResponse(
     private record FeedLogResponse(
             Long animalId,
             String feedType,
-            Integer feedAmount,
+            Float feedAmount,
             LocalDateTime feedDateTime,
             String significant
     ) {
